toolkit/components/extensions/test/mochitest/test_ext_notifications.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for notifications</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

add_task(function* test_notification() {
  function backgroundScript() {
    let opts = {
      type: "basic",
      title: "Testing Notification",
      message: "Carry on",
    };

    browser.notifications.create("", opts).then(id => {
      browser.test.sendMessage("running", id);
      browser.test.notifyPass("background test passed");
    });
  }

  let extension = ExtensionTestUtils.loadExtension({
    manifest: {
      permissions: ["notifications"],
    },
    background: `(${backgroundScript})()`,
  });
  yield extension.startup();
  let x = yield extension.awaitMessage("running");
  is(x, "0", "got correct id from notifications.create");
  yield extension.awaitFinish();
  yield extension.unload();
});

add_task(function* test_notification_on_closed() {
  function backgroundScript() {
    let opts = {
      type: "basic",
      title: "Testing Notification",
      message: "Carry on",
    };

    // Test onClosed listener.
    browser.notifications.onClosed.addListener(id => {
      browser.test.sendMessage("closed", id);
      browser.test.notifyPass("background test passed");
    });

    browser.notifications.create("5", opts).then(id => {
      return browser.notifications.create("5", opts);
    }).then(id => {
      browser.test.sendMessage("running", id);
    });
  }

  let extension = ExtensionTestUtils.loadExtension({
    manifest: {
      permissions: ["notifications"],
    },
    background: `(${backgroundScript})()`,
  });
  yield extension.startup();
  let x = yield extension.awaitMessage("closed");
  is(x, "5", "got correct id from onClosed listener");
  x = yield extension.awaitMessage("running");
  is(x, "5", "got correct id from notifications.create");
  yield extension.awaitFinish();
  yield extension.unload();
});

add_task(function* test_notification_clear() {
  function backgroundScript() {
    let opts = {
      type: "basic",
      title: "Testing Notification",
      message: "Carry on",
    };

    browser.notifications.onClosed.addListener(id => {
      browser.test.sendMessage("closed", id);
    });

    browser.notifications.create("99", opts).then(id => {
      return browser.notifications.clear(id);
    }).then(wasCleared => {
      browser.test.sendMessage("cleared", wasCleared);
      browser.test.notifyPass("background test passed");
    });
  }

  let extension = ExtensionTestUtils.loadExtension({
    manifest: {
      permissions: ["notifications"],
    },
    background: `(${backgroundScript})()`,
  });
  yield extension.startup();
  let x = yield extension.awaitMessage("closed");
  is(x, "99", "got correct id from onClosed listener");
  x = yield extension.awaitMessage("cleared");
  is(x, true, "got correct boolean from notifications.clear");
  yield extension.awaitFinish();
  yield extension.unload();
});

add_task(function* test_notifications_empty_getAll() {
  function backgroundScript() {
    browser.notifications.getAll().then(notifications => {
      browser.test.assertTrue(Array.isArray(notifications),
        "getAll() returned an array");
      browser.test.assertEq(notifications.length, 0, "the array was empty");
      browser.test.notifyPass("getAll empty");
    });
  }

  let extension = ExtensionTestUtils.loadExtension({
    manifest: {
      permissions: ["notifications"],
    },
    background: `(${backgroundScript})()`,
  });
  yield extension.startup();
  yield extension.awaitFinish("getAll empty");
  yield extension.unload();
});

add_task(function* test_notifications_populated_getAll() {
  function backgroundScript() {
    let opts = {
      type: "basic",
      title: "Testing Notification",
      message: "Carry on",
    };

    browser.notifications.create("p1", opts).then(() => {
      return browser.notifications.create("p2", opts);
    }).then(() => {
      return browser.notifications.getAll();
    }).then(notifications => {
      browser.test.assertTrue(Array.isArray(notifications),
        "getAll() returned an array");
      browser.test.assertEq(notifications.length, 2,
        "the array contained two notification ids");
      browser.test.assertTrue(notifications.includes("p1"),
        "the array contains the first notification");
      browser.test.assertTrue(notifications.includes("p2"),
        "the array contains the second notification");
      browser.test.notifyPass("getAll populated");
    });
  }

  let extension = ExtensionTestUtils.loadExtension({
    manifest: {
      permissions: ["notifications"],
    },
    background: `(${backgroundScript})()`,
  });
  yield extension.startup();
  yield extension.awaitFinish("getAll populated");
  yield extension.unload();
});

</script>

</body>
</html>
